requirements.txt
flask==2.0.3
flask-cors==3.0.10
gunicorn==20.1.0
gevent==21.12.0
requests==2.27.1
# prometheus-flask-exporter
pyaml
# matplotlib
# flask-wtf
# flask-login